UI 관련 기본 구현 추가 및 더미 일부 구현

+++ Assets/Scripts/System/BattleStat.cs
@@ -21,7 +21,7 @@ public class BattleStat
 	}
 	public bool IsDead => Hp == 0;
 	public int MaxHp { get; }
-	public float Speed { get; }
+	public int Speed { get; }
 	private List<IOption> optionList;
 	//field scope 기믹
 	private List<IBuff> buffList;
