+++ Assets/Scripts/System/ITurnObject.cs
@@ -4,5 +4,5 @@ public interface ITurnObject
 	//todo: 필요?
 	public IUpdatableRoutine UpdatableRoutine { get; }
 	public void StartTurn();
-	public float TurnSpeed { get; }
+	public int TurnSpeed { get; }
 }
